requirements.txt
fastapi
uvicorn[standard]
weaviate-client
openai
python-dotenv
pypdf
python-multipart
gradio
requests